Extract session capture/restore into WindowSessionMixin
Continue decomposing window.py: move capture_session, restore_session,
_restore_split_tab and _close_all_tabs verbatim into
sshpilot/window_session.py as WindowSessionMixin, inherited by MainWindow.
Pure code move — signatures and self. state access unchanged.

TerminalWidget is imported in the mixin the same way window.py imports it
(from .terminal import TerminalWidget); both bind to the one sshpilot.terminal
module in the same import cycle, so the isinstance() checks in capture/restore
use the identical class object the app and the session tests check against.

(cherry picked from commit 598da54ae82cba91add4b67539879e8cd8d8cd49)
